# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/chat/completion/stream/{session_id}": {
            "post": {
                "description": "流式输出聊天",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "流式输出聊天",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话 ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "用户输入及参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.CompletionStream.userInput"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/tue/course/create": {
            "post": {
                "description": "创建课程基础参数，绑定或创建题目、资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "创建课程",
                "parameters": [
                    {
                        "description": "课程内容",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Course"
                        }
                    }
                }
            }
        },
        "/tue/course/list": {
            "get": {
                "description": "获取课程列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "获取课程列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页参数",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_expr",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "start_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-entity_PaginatedTotalResponse-schema_Course"
                        }
                    }
                }
            }
        },
        "/tue/course/update": {
            "post": {
                "description": "更新课程基础参数，增量更新 题目、资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "更新课程",
                "parameters": [
                    {
                        "description": "课程内容",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Course"
                        }
                    }
                }
            }
        },
        "/tue/course/{id}": {
            "get": {
                "description": "获取单个课程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "获取单个课程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "课程 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-schema_Course"
                        }
                    }
                }
            }
        },
      "/tue/course/{id}/delete": {
            "post": {
                "description": "删除课程",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Course"
                ],
                "summary": "删除课程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "课程 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "$ref": "#/definitions/entity.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "chat.CompletionStream.userInput": {
            "type": "object",
            "required": [
                "model_name",
                "provider_name",
                "question"
            ],
            "properties": {
                "bot_id": {
                    "type": "integer"
                },
                "enable_context": {
                    "type": "boolean"
                },
                "model_name": {
                    "description": "Model.Name 准确的模型名称",
                    "type": "string"
                },
                "provider_name": {
                    "description": "Provider.Name 准确的供应商名称",
                    "type": "string"
                },
                "question": {
                    "type": "string"
                },
                "system_prompt": {
                    "description": "系统提示词",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-entity_PaginatedTotalResponse-schema_Course": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.PaginatedTotalResponse-schema_Course"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.CommonResponse-schema_Course": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "代码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Course"
                        }
                    ]
                },
                "msg": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
...
        "entity.PaginatedTotalResponse-schema_Course": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Course"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "schema.Course": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "课程描述",
                    "type": "string"
                },
                "exams": {
                    "description": "课程考试",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.CourseExam"
                    }
                },
                "id": {
                    "description": "原始数据",
                    "type": "integer"
                },
                "name": {
                    "description": "课程名称",
                    "type": "string"
                },
                "resources": {
                    "description": "关联数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.CourseResource"
                    }
                },
                "sorted_data": {
                    "description": "排好序的数据",
                    "type": "array",
                    "items": {}
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "schema.CourseExam": {
            "type": "object",
            "properties": {
                "course_id": {
                    "description": "关联课程ID",
                    "type": "integer"
                },
                "exam": {
                    "description": "考试详细信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Exam"
                        }
                    ]
                },
                "exam_id": {
                    "description": "关联考试ID",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "sort_order": {
                    "description": "考试排序",
                    "type": "integer"
                }
            }
        },
        "schema.CourseResource": {
            "type": "object",
            "properties": {
                "course_id": {
                    "description": "关联课程ID",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "resource": {
                    "description": "资源详细信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Resource"
                        }
                    ]
                },
                "resource_id": {
                    "description": "关联资源ID",
                    "type": "integer"
                },
                "sort_order": {
                    "description": "资源排序",
                    "type": "integer"
                }
            }
        },
        "schema.Exam": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "考试描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "limit_time": {
                    "description": "考试限时（单位：秒）",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "problems": {
                    "description": "考试包含的大题",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ExamProblem"
                    }
                },
                "subjects": {
                    "description": "所属科目分类",
                    "type": "string"
                },
                "total_score": {
                    "description": "考试总分（单位：0.01分）",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "schema.ExamProblem": {
            "type": "object",
            "properties": {
                "exam_id": {
                    "description": "关联考试ID",
                    "type": "integer"
                },
                "problem": {
                    "description": "题目详细信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.Problem"
                        }
                    ]
                },
                "problem_id": {
                    "description": "关联题目ID",
                    "type": "integer"
                },
                "score": {
                    "description": "题目分值（1表示0.01分）",
                    "type": "integer"
                },
                "sort_order": {
                    "description": "题目排序",
                    "type": "integer"
                }
            }
        },
...
        "schema.Problem": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "答案（JSON存储ProblemAnswer）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ProblemAnswer"
                        }
                    ]
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "支持HTML/Markdown",
                    "type": "string"
                },
                "difficulty": {
                    "description": "难度等级 1-5",
                  "type": "integer",
                  "maximum": 5,
                  "minimum": 1
                },
                "explanation": {
                    "description": "答案解析",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "options": {
                    "description": "选项（JSON存储ProblemOption数组）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ProblemOption"
                    }
                },
                "subject": {
                    "description": "所属科目/分类",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/schema.ProblemType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "schema.ProblemAnswer": {
            "type": "object",
            "properties": {
                "answer": {
                    "description": "选择题：存储正确选项ID []uint\n填空题：存储多个填空关键词 []string\n判断题：true/false\n简答题：文本答案 string"
                }
            }
        },
        "schema.ProblemOption": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "correct": {
                    "description": "是否正确答案",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "schema.ProblemType": {
            "type": "string",
            "enum": [
                "single_choice",
                "multiple_choice",
                "fill_blank",
                "short_answer",
                "true_false"
            ],
            "x-enum-varnames": [
                "SingleChoice",
                "MultipleChoice",
                "FillBlank",
                "ShortAnswer",
                "TrueFalse"
            ]
        },
...
        "schema.Resource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "资源描述",
                    "type": "string"
                },
                "file_key": {
                    "description": "文件的 uuid key",
                    "type": "string"
                },
                "file_name": {
                    "description": "OSS 中的文件名",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "origin_file_name": {
                    "description": "原始文件名",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}